+++ src/@types/IssuerSignature.ts
@@ -16,3 +16,13 @@ export interface SignedCredential {
   header: Record<string, any>
   issuer: string
 }
+
+export interface IssuerKeyJWK {
+  kty: string
+  crv: string
+  d: string
+  x: string
+  y: string
+  alg: string
+  use: string
+}
